Fix more errors 2

+++ tasks/seq/example/src/ops_seq.cpp
@@ -17,7 +17,7 @@ bool TestTaskSequential::validation() {
 
 bool TestTaskSequential::run() {
   internal_order_test();
-  for (size_t i = 0; i < input_; i++) {
+  for (int i = 0; i < input_; i++) {
     res++;
   }
   return true;
